pyproject.toml
[tool.poetry]
name = "openai-batch"
version = "0.1.0"
description = ""
authors = ["observer <wozluohd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.23.6"
pydantic = "^2.7.1"
python-daemon = "^3.0.1"
pytest = "^8.2.0"
sqlmodel = "^0.0.18"
tomli-w = "^1.0.0"
typer = "^0.12.3"
rich = "^13.7.1"
python-crontab = "^3.0.0"
python-pidfile = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
